+++ e2e/src/specs/milestone/milestoneView.spec.ts
@@ -4,7 +4,7 @@ import { milestoneList } from '../../pages/milestone/list.po';
 import { milestoneView } from '../../pages/milestone/view.po';
 import { ProjectHelper } from '../../helpers/project.helper';
 import { TestRun } from '../../../../src/app/shared/models/testrun';
-import { Milestone } from '../../../../src/app/shared/models/milestone';
+import { Milestone } from '../../../../src/app/shared/models/milestones/milestone';
 import usersTestData from '../../data/users.json';
 import using from 'jasmine-data-provider';
 import loginTestRunJson from '../../data/import/milestoneView/login.json';