+++ e2e/src/specs/testrun/view_manual.testrun.spec.ts
@@ -5,7 +5,7 @@ import { testrunList } from '../../pages/testrun/list.po';
 import { suiteView } from '../../pages/suite/view.po';
 import { ProjectHelper } from '../../helpers/project.helper';
 import { TestSuite } from '../../../../src/app/shared/models/test-suite';
-import { Milestone } from '../../../../src/app/shared/models/milestone';
+import { Milestone } from '../../../../src/app/shared/models/milestones/milestone';
 import { TestRun } from '../../../../src/app/shared/models/testrun';
 import users from '../../data/users.json';
 import using from 'jasmine-data-provider';