+++ src/app/shared/models/milestones/report-toggle.ts
@@ -1,0 +1,11 @@
+import {ReportOption} from './report-option';
+
+export  class ReportToggle {
+  option: ReportOption;
+  state: boolean;
+
+  constructor(option: ReportOption, state: boolean) {
+    this.option = option;
+    this.state = state;
+  }
+}